implement handlers

+++ bot/__init__.py
@@ -5,9 +5,11 @@ from discord.ext import commands
 from matchmaker import Database, MatchMaker, Config
 from matchmaker.tables import Round
 from matchmaker.template import ColumnQuery, QueryKind, Max
+from matchmaker.event import EventKind
 
 from .config import BotConfig
 from .cogs import MatchMakerCog, DatabaseCog, AdminCog
+from .handlers import MatchStartHandler, MatchEndHandler, ResultHandler
 
 __all__ = ("MatchMakerBot", "Database", "MatchMaker")
 
@@ -31,6 +33,9 @@ class MatchMakerBot(commands.Bot):
         round_id = 0 if round_id is None else round_id
 
         self.mm = MatchMaker(mmcfg, Round(round_id=round_id + 1))
+        self.mm.register_handler(MatchStartHandler(self.loop))
+        self.mm.register_handler(MatchEndHandler(self.loop))
+        self.mm.register_handler(ResultHandler(self.db))
 
         for cog in COGS:
             self.add_cog(cog(self.db, self.mm))
@@ -43,8 +48,29 @@ class MatchMakerBot(commands.Bot):
 
     async def on_message(self, message):
         is_command = message.content[0] == self.command_prefix
-        if message.channel.name != self.config.channel:
+        if is_command and message.channel.name != self.config.channel:
             await message.delete()
+        elif is_command and message.content.startswith("+queue"):
+            try:
+                handler_index = self.mm.evmap[EventKind.ROUND_START].index(
+                    MatchStartHandler(self.loop)
+                )
+                handler = self.mm.evmap[EventKind.ROUND_START][handler_index]
+                assert isinstance(handler, MatchStartHandler)
+                handler.channel = message.channel
+            except ValueError:
+                pass
+        elif is_command and message.content.startswith("+result"):
+            try:
+                handler_index = self.mm.evmap[EventKind.ROUND_END].index(
+                    MatchEndHandler(self.loop)
+                )
+                handler = self.mm.evmap[EventKind.ROUND_END][handler_index]
+                assert isinstance(handler, MatchEndHandler)
+                handler.channel = message.channel
+            except ValueError:
+                pass
+
         await super().on_message(message)
 
     async def on_ready(self):
